PLacing test in the test module
