+++ crates/cloud-wallet-openid4vc/src/formats/mdoc/tests/mod.rs
@@ -1,4 +1,5 @@
 mod annex_d;
+mod claims;
 mod parser;
 mod revocation;
 mod verifier;
